requirements.txt
# Core dependencies
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# LLM Providers (install one or both)
anthropic>=0.18.0
openai>=1.12.0

# Scheduling
APScheduler>=3.10.0

# Caching (optional - for Redis support)
redis>=5.0.0

# Production server (optional)
gunicorn>=21.0.0
